requirements.txt
--find-links https://download.pytorch.org/whl/torch_stable.html
numpy>=1.18
torch==1.10.1+cu111
mujoco-py<2.2,>=2.1
gym
tensorboardX==1.2
pyyaml==3.12
gitpython==3.0.8
imageio==2.14.1
